Upgrade to LLVM 6 and 7

+++ test/test086.c
@@ -18,9 +18,9 @@ int main() {
   char inc0 = 0, inc1 = 0;
   int diff = 0, a = 100;
   int *p = &a;
-  inc0 =(int)p;
+  inc0 =(long)p;
   ++(*p++);  //++(*p++);
-  inc1 =(int)p;
+  inc1 =(long)p;
   diff =  inc1-inc0;
   diff += 2;
   return diff;
